requirements.txt
# Core Framework
Flask==3.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Werkzeug==3.0.1
python-dotenv==1.0.1
Jinja2==3.1.3
click==8.1.7
itsdangerous==2.1.2
blinker==1.7.0

# Database and ORM
SQLAlchemy==2.0.25
Flask-SQLAlchemy==3.1.1
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
PyJWT==2.8.0

# OpenAI Integration
openai==1.12.0
aiohttp==3.9.1

# Vector Database
pgvector==0.2.4

# Utilities
python-dateutil==2.8.2
requests==2.31.0
validators==0.22.0

# Development Tools
pytest==7.4.4
black==24.1.1
flake8==7.0.0
python-dotenv==1.0.1

# Additional requirements
WTForms==3.1.2
email-validator==2.1.0.post1
chromadb==0.4.22
bcrypt==4.1.2 